requirements.txt
streamlit
pandas
pillow
matplotlib
fpdf
requests
openpyxl
